Address build shortcomings
