+++ src/flatbuffers/schema/genai_lora.fbs.h
@@ -1,5 +1,6 @@
 // automatically generated by the FlatBuffers compiler, do not modify
 
+// clang-format off
 
 #ifndef FLATBUFFERS_GENERATED_GENAILORA_GENERATORS_LORA_PARAMETERS_H_
 #define FLATBUFFERS_GENERATED_GENAILORA_GENERATORS_LORA_PARAMETERS_H_
