+++ tools/python/util/__init__.py
@@ -2,7 +2,6 @@
 # Licensed under the MIT License.
 
 from .logger import get_logger
-from .lora_helpers import save_lora_params_to_flatbuffers, add_adapters_to_genai_config
 from .run import run
 from .android import *
 from .platform_helpers import is_linux, is_mac, is_windows
